mqtt-streaming: fix flaky 'duplicate publish given two client connections' test
It appears the flakiness was caused by the two connections using the same
connection id for different connections in the same session. I'm not intimately
familiar with this codebase, but AFAICS it seems this connection id is an
internal concern to the component, and not an MQTT concept. I think we should
probably provide an API that does not bother the user with this concept at
all, and manages non-overlapping connection ids entirely internal to the
component. I don't get the impression that it was the purpose of the test to
test the behaviour in case of overlapping connection ids.

Fixes #454

After this change the 'unsubscribe a client session' seems to fail
a lot on CI, but that seems unrelated and something to fix in a separate PR.