fix: don't use overlapping session ids in MqttFlowTest
While looking into #468, I noticed the two failing tests were
sharing the same session id, which reminded of #456.

While in this case the two tests aren't sharing the same
session, and I haven't investigated the details of this
codebase further, I'm curious to see if the problem remains
when we use unique session ids.